When dispatching routes to a handler, use the absolute path not the full URI
Note that using the absolute path allows query strings in the URI to resolve to
the expected resource rather than returning a 404 error. Previously the only
way to allow a query string was to use the wildcard as the last atom of the
route spec, but this has the potentially undesirable consequence of accepting
more URIs than expected.